requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==4.1.7
django-taggit==3.0.0
Markdown==3.4.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
python-decouple==3.8
sqlparse==0.4.3
tomli==2.0.1
